Get Balances for each staking contract on oportunities screen
